refactor: 프로젝트 폴더 이름 변경 (#90)
