+++ src/features/project-d/pages/ProjectDetailPage.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from "react";
 import { useParams, useNavigate } from "react-router-dom";
 import * as S from "./ProjectDetailPage.styled";
 import api from "@/api/axios";
-import { getPosts } from "@/features/project/services/postService";
+import { getPosts } from "@/features/project-d/services/postService";
 
 interface UserCompanyResponse {
   id: number;
